third_party/WebKit/LayoutTests/http/tests/media/video-play-stall-expected.txt

Test that stalled, timeupdate and waiting events are sent when media load stalls in the middle.

RUN(video.play())
EVENT(durationchange)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
EVENT(timeupdate)
EVENT(stalled)
END OF TEST
